# Log: check severity before backend