--- a/src/log/log.cr
+++ b/src/log/log.cr
@@ -44,9 +44,10 @@ class Log
                              } %}
     # Logs a message if the logger's current severity is lower or equal to `{{severity}}`.
     def {{method.id}}(*, exception : Exception? = nil)
-      return unless backend = @backend
       severity = Severity.new({{severity}})
       return unless level <= severity
+
+      return unless backend = @backend
 
       dsl = Emitter.new(@source, severity, exception)
       result = yield dsl