@bcardiff 2020-06-01
Only if you have more Logs with backend 😉 . 
But since now the default source filter is `*`, instead of top-level as before, this is a good change.